Added Menus
Added menus to main menu and chore item view
Added Contextual Action Menu when long-clicking items
Fixed long-click transition animation bug which was causing a crash
Simplified some layout files by removing separate "content" files

+++ app/src/main/java/com/taylor/chorelist/ChoreDatabaseHelper.java
@@ -28,7 +28,7 @@ public class ChoreDatabaseHelper {
         return chore_list;
     }
 
-    static public ChoreItem pull_chore_item(Context context, int index) {
+    static public ChoreItem pull_chore_item(Context context, long index) {
         //Get a database
         SQLiteDatabase db = ChoreItemDatabase.get_readable_db(context);
         //Create our SQL query (ID = index passed)
